backend/requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
python-multipart
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
pillow==11.0.0
sympy==1.13.1
torch
typing_extensions==4.9.0
transformers